reorganize scenario/utils
